Evaluates Statements of An Expression Language
The statements of that expression language consist of an arithmetic expression followed by a list
of assignments. Assignments are separated from the expression and each other by commas. A
semicolon terminates the expression. The arithmetic expressions are fully parenthesized infix
expressions containing integer literals and variables. The syntax of a single input file line is
described grammar below:
statement → expression ',' assignments ';'
expression → '(' operand operator operand ')'
operator → '+' | '
-
'
operand → literal | variable | expression
assignments → assignments ',' assignment | assignment
assignment → variable '=' literal

The program reads in the arithmetic expression and encodes the expression as a binary tree. After
the expression has been read in, the variable assignments are read in and the variables and their
values of the variables are placed into the symbol table. Finally the expression is evaluated
recursively.